requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.4
pydantic-settings==2.1.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
psycopg2-binary
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
passlib[bcrypt]==1.7.4

bcrypt==3.2.2
pytest==9.0.1
httpx==0.27.0
PyJWT==2.10.1

google-generativeai==0.8.5


alembic==1.13.3
httpx==0.27.0


pytest-mock

# Rate limiting
slowapi==0.1.9

# AWS integration
boto3==1.35.0
botocore==1.35.0

# Structured logging for CloudWatch
python-json-logger==2.0.7
